Implement locking for scheduler jobs

+++ SchedulerBot.Client/Scheduler/EventDeleteJob.cs
@@ -1,6 +1,9 @@
 ﻿using System;
 using System.Threading.Tasks;
+using DSharpPlus;
 using Quartz;
+using RedLockNet;
+using SchedulerBot.Client.Exceptions;
 using SchedulerBot.Data.Services;
 
 namespace SchedulerBot.Client.Scheduler
@@ -10,10 +13,25 @@ namespace SchedulerBot.Client.Scheduler
         public async Task Execute(IJobExecutionContext context)
         {
             JobDataMap jobDataMap = context.MergedJobDataMap;
+            DiscordClient client = (DiscordClient)jobDataMap["client"];
             Guid eventId = (Guid)jobDataMap["eventId"];
+            ulong channelId = (ulong)jobDataMap["channelId"];
             IEventService eventService = (IEventService)jobDataMap["eventService"];
+            IDistributedLockFactory redlockFactory = (IDistributedLockFactory)jobDataMap["redlockFactory"];
 
-            await eventService.DeleteEventAsync(eventId);
+            var guildId = (await client.GetChannelAsync(channelId)).GuildId;
+
+            using (var redlock = await redlockFactory.CreateLockAsync(guildId.ToString(), TimeSpan.FromSeconds(15), TimeSpan.FromSeconds(10), TimeSpan.FromSeconds(0.5)))
+            {
+                if (redlock.IsAcquired)
+                {
+                    await eventService.DeleteEventAsync(eventId);
+                }
+                else
+                {
+                    throw new RedisLockAcquireException($"Cannot acquire lock for guild {guildId}");
+                }
+            }
         }
     }
 }
